move exception to common package
