+++ actuator/src/main/java/org/tron/core/actuator/MarketCancelOrderActuator.java
@@ -43,6 +43,7 @@ import org.tron.protos.Protocol.Transaction.Result.code;
 import org.tron.protos.contract.AssetIssueContractOuterClass.AssetIssueContract;
 import org.tron.protos.contract.MarketContract.MarketCancelOrderContract;
 
+import static org.tron.core.actuator.ActuatorConstant.CONTRACT_NOT_EXIST;
 import static org.tron.core.actuator.ActuatorConstant.TX_RESULT_NULL;
 
 @Slf4j(topic = "actuator")
@@ -147,7 +148,7 @@ public class MarketCancelOrderActuator extends AbstractActuator {
   @Override
   public boolean validate() throws ContractValidateException {
     if (this.any == null) {
-      throw new ContractValidateException("No contract!");
+      throw new ContractValidateException(CONTRACT_NOT_EXIST);
     }
     if (chainBaseManager == null) {
       throw new ContractValidateException("No account store or dynamic store!");
